VIZ: unify the line and heatmap dendrogram styles in a single class
